add table_schema and table_name in search

+++ build.sql
@@ -1612,14 +1612,17 @@ $$;
 create function schema.search(
     _schema text = null,
     _type text = null,
-    _search text = null
+    _search text = null,
+    _verbose boolean = true
 )
 returns table (
     type text,
     schema text,
     name text,
     comment text,
-    definition text
+    definition text,
+    table_schema text,
+    table_name text
 )
 language plpgsql
 as 
@@ -1635,6 +1638,12 @@ begin
     if _schemas is null or _schemas = '{}' then
         raise exception 'No schema found for expression: %', _schema;
     end if;
+
+    if _verbose then
+        raise notice 'Search schema: %', _schemas;
+        raise notice 'Search type: %', _type;
+        raise notice 'Search text: %', _search;
+    end if;
     
     if schema._temp_exists('search') then
         drop table pg_temp.search;
@@ -1646,7 +1655,9 @@ begin
         t.schema,
         t.name,
         t.comment,
-        t.definition
+        t.definition,
+        null as table_schema,
+        null as table_name
     from schema._extensions() t
     where
         schema._search_filter(t, _type, _search);
@@ -1657,18 +1668,22 @@ begin
         t.schema,
         t.name,
         t.comment,
-        t.definition
+        t.definition,
+        null as table_schema,
+        null as table_name
     from schema._types(_schemas) t
     where
         schema._search_filter(t, _type, _search);
         
     insert into pg_temp.search
-    select  
+    select 
         t.type,
         t.schema,
         t.name,
         t.comment,
-        t.definition
+        t.definition,
+        null as table_schema,
+        null as table_name
     from schema._enums(_schemas) t
     where
         schema._search_filter(t, _type, _search);
@@ -1679,7 +1694,9 @@ begin
         t.schema,
         t.name,
         t.comment,
-        t.definition
+        t.definition,
+        null as table_schema,
+        null as table_name
     from schema._ranges(_schemas) t
     where
         schema._search_filter(t, _type, _search);
@@ -1690,7 +1707,9 @@ begin
         t.schema,
         t.name,
         t.comment,
-        t.definition
+        t.definition,
+        null as table_schema,
+        null as table_name
     from schema._domains(_schemas) t
     where
         schema._search_filter(t, _type, _search);
@@ -1703,7 +1722,9 @@ begin
         t.schema,
         t.name,
         t.comment,
-        t.definition
+        t.definition,
+        null as table_schema,
+        null as table_name
     from schema._tables_full(_schemas) t
     where
         schema._search_filter(t, _type, _search);
@@ -1714,14 +1735,18 @@ begin
         sub.schema,
         sub.name,
         sub.comment,
-        sub.definition
+        sub.definition,
+        sub.table_schema,
+        sub.table_name
     from (
         select
             t.type, 
             t.schema, 
             t.table_name || '.' || quote_ident(t.name) as name, 
             t.comment,
-            t.definition
+            t.definition,
+            t.table_schema,
+            t.table_name
         from _columns t
     ) sub
     where
@@ -1733,7 +1758,9 @@ begin
         t.schema, 
         t.name, 
         t.comment,
-        t.definition
+        t.definition,
+        t.table_schema,
+        t.table_name
     from _constraints t
     where
         schema._search_filter(t, _type, _search);
@@ -1744,7 +1771,9 @@ begin
         t.schema, 
         t.name, 
         t.comment,
-        t.definition
+        t.definition,
+        t.table_schema,
+        t.table_name
     from _indexes t
     where
         schema._search_filter(t, _type, _search);
@@ -1755,7 +1784,9 @@ begin
         t.schema, 
         t.name, 
         t.comment,
-        t.definition
+        t.definition,
+        t.table_schema,
+        t.table_name
     from _triggers t
     where
         schema._search_filter(t, _type, _search);
@@ -1766,7 +1797,9 @@ begin
         t.schema, 
         t.name, 
         t.comment,
-        t.definition
+        t.definition,
+        t.table_schema,
+        t.table_name
     from _policies t
     where
         schema._search_filter(t, _type, _search);
@@ -1777,20 +1810,26 @@ begin
         sub.schema,
         sub.name,
         sub.comment,
-        sub.definition
+        sub.definition,
+        sub.table_schema,
+        sub.table_name
     from (
         select 
             t.type,
             t.schema,
             t.name,
             t.comment,
-            string_agg(t.definition, E'\n') as definition
+            string_agg(t.definition, E'\n') as definition,
+            t.table_schema,
+            t.table_name
         from _sequences t
         group by
             t.type,
             t.schema,
             t.name,
-            t.comment
+            t.comment,
+            t.table_schema,
+            t.table_name
     ) sub
     where
         schema._search_filter(sub, _type, _search);
@@ -1801,7 +1840,9 @@ begin
         t.schema,
         t.name,
         t.comment,
-        t.definition
+        t.definition,
+        null as table_schema,
+        null as table_name
     from schema._views(_schemas) t
     where
         schema._search_filter(t, _type, _search);
@@ -1812,7 +1853,9 @@ begin
         t.schema,
         t.name,
         t.comment,
-        t.definition
+        t.definition,
+        null as table_schema,
+        null as table_name
     from schema._routines(_schemas) t
     where
         schema._search_filter(t, _type, _search);
@@ -1823,7 +1866,9 @@ begin
         t.schema,
         t.name,
         t.comment,
-        t.definition
+        t.definition,
+        null as table_schema,
+        null as table_name
     from schema._rules(_schemas) t
     where
         schema._search_filter(t, _type, _search);
@@ -1834,21 +1879,23 @@ begin
         t.schema,
         t.name,
         t.comment,
-        t.definition
+        t.definition,
+        null as table_schema,
+        null as table_name
     from schema._aggregates(_schemas) t
     where
         schema._search_filter(t, _type, _search);
     
     return query
     select 
-        r.type, r.schema, r.name, r.comment, r.definition
+        r.type, r.schema, r.name, r.comment, r.definition, r.table_schema, r.table_name
     from pg_temp.search r
     order by 
         r.schema, r.type desc, r.name;
 end;
 $$;
 
-comment on function schema.search(text, text, text) is 'Search and retrieve the schema objects in the database.';
+comment on function schema.search(text, text, text, boolean) is 'Search and retrieve the schema objects in the database.';
 /* #endregion search */
 
 /* #region dump */
