+++ src/search.sql
@@ -1,14 +1,17 @@
 create function schema.search(
     _schema text = null,
     _type text = null,
-    _search text = null
+    _search text = null,
+    _verbose boolean = true
 )
 returns table (
     type text,
     schema text,
     name text,
     comment text,
-    definition text
+    definition text,
+    table_schema text,
+    table_name text
 )
 language plpgsql
 as 
@@ -24,6 +27,12 @@ begin
     if _schemas is null or _schemas = '{}' then
         raise exception 'No schema found for expression: %', _schema;
     end if;
+
+    if _verbose then
+        raise notice 'Search schema: %', _schemas;
+        raise notice 'Search type: %', _type;
+        raise notice 'Search text: %', _search;
+    end if;
     
     if schema._temp_exists('search') then
         drop table pg_temp.search;
@@ -35,7 +44,9 @@ begin
         t.schema,
         t.name,
         t.comment,
-        t.definition
+        t.definition,
+        null as table_schema,
+        null as table_name
     from schema._extensions() t
     where
         schema._search_filter(t, _type, _search);
@@ -46,18 +57,22 @@ begin
         t.schema,
         t.name,
         t.comment,
-        t.definition
+        t.definition,
+        null as table_schema,
+        null as table_name
     from schema._types(_schemas) t
     where
         schema._search_filter(t, _type, _search);
         
     insert into pg_temp.search
-    select  
+    select 
         t.type,
         t.schema,
         t.name,
         t.comment,
-        t.definition
+        t.definition,
+        null as table_schema,
+        null as table_name
     from schema._enums(_schemas) t
     where
         schema._search_filter(t, _type, _search);
@@ -68,7 +83,9 @@ begin
         t.schema,
         t.name,
         t.comment,
-        t.definition
+        t.definition,
+        null as table_schema,
+        null as table_name
     from schema._ranges(_schemas) t
     where
         schema._search_filter(t, _type, _search);
@@ -79,7 +96,9 @@ begin
         t.schema,
         t.name,
         t.comment,
-        t.definition
+        t.definition,
+        null as table_schema,
+        null as table_name
     from schema._domains(_schemas) t
     where
         schema._search_filter(t, _type, _search);
@@ -92,7 +111,9 @@ begin
         t.schema,
         t.name,
         t.comment,
-        t.definition
+        t.definition,
+        null as table_schema,
+        null as table_name
     from schema._tables_full(_schemas) t
     where
         schema._search_filter(t, _type, _search);
@@ -103,14 +124,18 @@ begin
         sub.schema,
         sub.name,
         sub.comment,
-        sub.definition
+        sub.definition,
+        sub.table_schema,
+        sub.table_name
     from (
         select
             t.type, 
             t.schema, 
             t.table_name || '.' || quote_ident(t.name) as name, 
             t.comment,
-            t.definition
+            t.definition,
+            t.table_schema,
+            t.table_name
         from _columns t
     ) sub
     where
@@ -122,7 +147,9 @@ begin
         t.schema, 
         t.name, 
         t.comment,
-        t.definition
+        t.definition,
+        t.table_schema,
+        t.table_name
     from _constraints t
     where
         schema._search_filter(t, _type, _search);
@@ -133,7 +160,9 @@ begin
         t.schema, 
         t.name, 
         t.comment,
-        t.definition
+        t.definition,
+        t.table_schema,
+        t.table_name
     from _indexes t
     where
         schema._search_filter(t, _type, _search);
@@ -144,7 +173,9 @@ begin
         t.schema, 
         t.name, 
         t.comment,
-        t.definition
+        t.definition,
+        t.table_schema,
+        t.table_name
     from _triggers t
     where
         schema._search_filter(t, _type, _search);
@@ -155,7 +186,9 @@ begin
         t.schema, 
         t.name, 
         t.comment,
-        t.definition
+        t.definition,
+        t.table_schema,
+        t.table_name
     from _policies t
     where
         schema._search_filter(t, _type, _search);
@@ -166,20 +199,26 @@ begin
         sub.schema,
         sub.name,
         sub.comment,
-        sub.definition
+        sub.definition,
+        sub.table_schema,
+        sub.table_name
     from (
         select 
             t.type,
             t.schema,
             t.name,
             t.comment,
-            string_agg(t.definition, E'\n') as definition
+            string_agg(t.definition, E'\n') as definition,
+            t.table_schema,
+            t.table_name
         from _sequences t
         group by
             t.type,
             t.schema,
             t.name,
-            t.comment
+            t.comment,
+            t.table_schema,
+            t.table_name
     ) sub
     where
         schema._search_filter(sub, _type, _search);
@@ -190,7 +229,9 @@ begin
         t.schema,
         t.name,
         t.comment,
-        t.definition
+        t.definition,
+        null as table_schema,
+        null as table_name
     from schema._views(_schemas) t
     where
         schema._search_filter(t, _type, _search);
@@ -201,7 +242,9 @@ begin
         t.schema,
         t.name,
         t.comment,
-        t.definition
+        t.definition,
+        null as table_schema,
+        null as table_name
     from schema._routines(_schemas) t
     where
         schema._search_filter(t, _type, _search);
@@ -212,7 +255,9 @@ begin
         t.schema,
         t.name,
         t.comment,
-        t.definition
+        t.definition,
+        null as table_schema,
+        null as table_name
     from schema._rules(_schemas) t
     where
         schema._search_filter(t, _type, _search);
@@ -223,18 +268,20 @@ begin
         t.schema,
         t.name,
         t.comment,
-        t.definition
+        t.definition,
+        null as table_schema,
+        null as table_name
     from schema._aggregates(_schemas) t
     where
         schema._search_filter(t, _type, _search);
     
     return query
     select 
-        r.type, r.schema, r.name, r.comment, r.definition
+        r.type, r.schema, r.name, r.comment, r.definition, r.table_schema, r.table_name
     from pg_temp.search r
     order by 
         r.schema, r.type desc, r.name;
 end;
 $$;
 
-comment on function schema.search(text, text, text) is 'Search and retrieve the schema objects in the database.';
+comment on function schema.search(text, text, text, boolean) is 'Search and retrieve the schema objects in the database.';
